co-locate theme hook with ThemeContextProvider
